src/requirements.txt
pyspark==3.5.0
psycopg2-binary
pandas